requirements.txt
sip>=6.10.0
pillow>=11.1.0